move OpenID4VCI code to own file
