+++ auth/api/iam/openid4vp.go
@@ -322,7 +322,7 @@ func (r Wrapper) handleAuthorizeRequestFromVerifier(ctx context.Context, tenantD
 		privateKey, err = userSession.Wallet.Key()
 		walletDID = userSession.Wallet.DID
 		targetWallet = holder.NewMemoryWallet(
-			r.JSONLDManager.DocumentLoader(),
+			r.jsonldManager.DocumentLoader(),
 			resolver.DIDKeyResolver{Resolver: didjwk.NewResolver()},
 			crypto.MemoryJWTSigner{Key: privateKey},
 			map[did.DID][]vc.VerifiableCredential{userSession.Wallet.DID: userSession.Wallet.Credentials},
